% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\encoding{UTF-8}
\name{simulate}
\alias{simulate}
\title{Simulation scenario used in the paper}
\usage{
simulate(n, d, eps = 0.5)
}
\arguments{
\item{n}{number of agents}

\item{d}{number of items; should be a multiple of the number of agents}

\item{eps}{parameter governing the dependence between the agents' valuations}
}
\description{
simulate standardized valuation matrices
}
\references{
Cornilly, D., Puccetti, G., Rüschendorf, L., & Vanduffel, S. (2020). 
New algorithms for fair allocation of indivisible goods with minimum inequality or minimum envy
criteria.
}
\author{
Dries Cornilly
}
\concept{simulate}
